Enhance logging configuration with JSON parsing and CLI options
- Add logger configuration parsing from JSON files
- Implement log level and group configuration via command-line arguments
- Update main application to use logger for output instead of direct console printing
- Improve logger flexibility with more configuration options
- Add support for log file and console logging configuration

+++ cpp/src/evaluation.cpp
@@ -39,7 +39,7 @@ double calculateScore(const uint8_t board[4][4]) {
         }
     }
 
-    return score;
+    return static_cast<double>(score);
 }
 
 // Find the maximum tile value
